feat: Add Vite environment declaration file and Firebase configuration
This commit adds the Vite environment declaration file (.firebaserc) and the Firebase configuration file (firebase.json). The .firebaserc file sets the default project to "portfolio-lokesh-sahukar", while the firebase.json file configures the hosting settings for the project. The hosting settings specify the public directory as "dist", ignore certain files and directories, and define a rewrite rule to redirect all requests to the index.html file. These changes are necessary for deploying the project to Firebase.